% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_manager_api.R
\name{rspm_api}
\alias{rspm_api}
\alias{rspm_get_repos}
\alias{rspm_get_packages}
\alias{rspm_get_package_sysreqs}
\alias{rspm_status}
\title{API to interact with RStudio package manager}
\usage{
rspm_get_repos(type = "R")

rspm_get_packages(id = 1L, limit = 10L, page, name, name_search)

rspm_get_package_sysreqs(package_name, id = 1L)

rspm_status()
}
\arguments{
\item{type}{character(1) value can only be 'R' in present API.}

\item{id}{character(1)}

\item{limit}{character(1)}

\item{page}{character(1)}

\item{name}{character(1)}

\item{name_search}{character(1)}

\item{character(1)}{package_name}
}
\value{






tibble status of RSPM
}
\description{
These functions provide an interface with the RStudio
package manager API as functions with the prefix 'rspm'.
URL: https://packagemanager.rstudio.com/\strong{api}/swagger/index.html to
the RSPM API swagger page.

Get available repos from RStudio Package Manager.



Fetch system requirements by package

Get the status of the Rstudio package manager
}
\examples{




}
